refactor: update ThemeToggle import paths and usage
