--- examples/common_nestjs_remix/apps/web/app/components/SheetMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SheetMenu";
