--- examples/common_nestjs_remix/apps/web/app/components/ThemeToggle/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ThemeToggle";
