--- examples/common_nestjs_remix/apps/web/app/components/index.ts
@@ -1,2 +1,0 @@
-export { default as SheetMenu } from "./SheetMenu";
-export { default as ThemeToggle } from "./ThemeToggle";
