+++ examples/common_nestjs_remix/apps/web/app/modules/Landing/Landing.page.tsx
@@ -1,6 +1,6 @@
 import type { MetaFunction } from "@remix-run/node";
 import { Link, useLoaderData } from "@remix-run/react";
-import { SheetMenu } from "~/components";
+import SheetMenu from "~/components/SheetMenu/SheetMenu";
 import { useLandingStore } from "~/modules/Landing/landingStore";
 
 export const meta: MetaFunction = () => {
